+++ surpyval/univariate/parametric/distributions/gumbel.py
@@ -342,7 +342,7 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
         """
         return np.log(sigma) + euler_gamma + 1
 
-    def mpp_x_transform(self, x: Numeric) -> Boxable:
+    def mpp_x_transform(self, x: npt.NDArray) -> Boxable:
         return x
 
     def mpp_y_transform(self, y: npt.NDArray, *params: Boxable) -> Boxable:
@@ -352,7 +352,7 @@ class Gumbel_(OptimisedFitMixin, ParametricFitter):
         out[mask] = np.nan
         return out
 
-    def mpp_inv_y_transform(self, y: Numeric, *params: Boxable) -> Boxable:
+    def mpp_inv_y_transform(self, y: npt.NDArray, *params: Boxable) -> Boxable:
         return 1 - np.exp(-np.exp(y))
 
     def unpack_rr(