+++ surpyval/univariate/parametric/distributions/rayleigh.py
@@ -394,7 +394,7 @@ class Rayleigh_(OptimisedFitMixin, ParametricFitter):
 
             return {"params": params}
 
-    def mpp_x_transform(self, x: Numeric) -> Boxable:
+    def mpp_x_transform(self, x: npt.NDArray) -> Boxable:
         return x
 
     def mpp_y_transform(self, y: npt.NDArray, *params: Boxable) -> Boxable:
@@ -404,7 +404,7 @@ class Rayleigh_(OptimisedFitMixin, ParametricFitter):
         out[mask] = np.nan
         return out
 
-    def mpp_inv_y_transform(self, y: Numeric, *params: Boxable) -> Boxable:
+    def mpp_inv_y_transform(self, y: npt.NDArray, *params: Boxable) -> Boxable:
         return 1 - np.exp(-(y**2))